+++ lib/src/televerse/models/models.dart
@@ -10,4 +10,3 @@ part 'webhook_exception.dart';
 part 'http_exception.dart';
 part 'long_polling_exception.dart';
 part 'televerse_exception.dart';
-part 'message_or_boolean.dart';